% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{learnItemEmb}
\alias{learnItemEmb}
\title{Title}
\usage{
learnItemEmb(
  train_data_path,
  emb_size,
  regularization,
  learning_rate,
  neg_sample_cnt,
  epoch
)
}
\arguments{
\item{epoch}{}
}
\value{

}
\description{
Title
}
