% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bestSubset}
\alias{bestSubset}
\title{Title}
\usage{
bestSubset(item_representations, ratings, n)
}
\arguments{
\item{item_representations}{Matrix}

\item{ratings}{Matrix}

\item{n}{int}
}
\value{
DataFrame with two columns "id" & "gain"
}
\description{
Title
}
\examples{
library(rDppDiversity)
bestSubset(matrix(c(-0.1, -0.15, 0.2, 0.3, -0.2, -0.3), nrow=3, ncol=2, byrow=TRUE), c(1, 2, 3), 3)
}
