% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bestSubset}
\alias{bestSubset}
\title{Given item set, item representation vector, and item ratings, find a subset with better relevance-diversity trade-off.}
\usage{
bestSubset(item_representations, ratings, n)
}
\arguments{
\item{item_representations}{matrix with nrow = num_items and ncol = feature_dim}

\item{ratings}{matrix with nrow = num_items and ncol = 1}

\item{n}{int. subset size}
}
\value{
dataframe with 2 columns: id & gain
}
\description{
Given item set, item representation vector, and item ratings, find a subset with better relevance-diversity trade-off.
}
\examples{
library(rDppDiversity)
bestSubset(matrix(c(-0.1, -0.15, 0.2, 0.3, -0.2, -0.3), nrow=3, ncol=2, byrow=TRUE), c(1, 2, 3), 3)
}
