% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{learnItemEmb}
\alias{learnItemEmb}
\title{Machine learning algorithm to learn item representations maximizing log likelihood under DPP assumption.}
\usage{
learnItemEmb(
  train_data_path,
  emb_size,
  regularization,
  learning_rate,
  neg_sample_cnt,
  epoch
)
}
\arguments{
\item{train_data_path}{string}

\item{emb_size}{int}

\item{regularization}{float}

\item{learning_rate}{float}

\item{neg_sample_cnt}{int}

\item{epoch}{int}
}
\value{
List contains: 1) learned item embedding; 2) item names; 3) log likelihood on each training step.
}
\description{
Machine learning algorithm to learn item representations maximizing log likelihood under DPP assumption.
}
\examples{
library(rDppDiversity)
data_path=system.file("extdata", "data.txt", package = "rDppDiversity")
learnItemEmb(train_data_path=data_path, emb_size=3, regularization=0.1, learning_rate=0.01, neg_sample_cnt=0, epoch=10)
}
